+++ test/psqlite_test.dart
@@ -1,9 +1,11 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:psqlite/psqlite.dart';
 import 'package:sqflite/sqflite.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 
 import '../example/lib/model/user.dart';
 import '../example/lib/services/user_storage_service.dart';
+
 
 Future main() async {
   late UserStorageService storageService;
@@ -99,5 +101,47 @@ Future main() async {
     await storageService.removeAll();
     final listOfUsers = await storageService.getListOfUsers();
     expect(listOfUsers.length, 0);
+  });
+
+  test('Remove users with filters', () async {
+    final user1 = User("1", "Liam", "Neeson", 12);
+    final user2 = User("2", "Mark", "Neeson", 13);
+    final user3= User("3", "John", "Neeson", 14);
+    final user4 = User("4", "Amy", "Neeson", 18);
+    final user5 = User("5", "Harald", "Neeson", 22);
+    await storageService.addUser(user1);
+    await storageService.addUser(user2);
+    await storageService.addUser(user3);
+    await storageService.addUser(user4);
+    await storageService.addUser(user5);
+    final filters = [FilterDb(UserColumnName.age.name, 18, ConditionDb.less)];
+    await storageService.removeUsers(filters);
+    final listOfUsers = await storageService.getListOfUsers();
+    expect(listOfUsers.length, 2);
+  });
+
+  test('Get users with filters', () async {
+    final user1 = User("1", "Liam", "Neeson", 12);
+    final user2 = User("2", "Mark", "Neeson", 13);
+    final user3= User("3", "John", "Neeson", 14);
+    final user4 = User("4", "Amy", "Neeson", 18);
+    final user5 = User("5", "Harald", "Neeson", 22);
+
+    await storageService.addUser(user1);
+    await storageService.addUser(user2);
+    await storageService.addUser(user3);
+    await storageService.addUser(user4);
+    await storageService.addUser(user5);
+
+    List<User> listOfUsers = await storageService.getListOfUsers();
+    expect(listOfUsers.length, 5);
+
+    final filters = [FilterDb(UserColumnName.age.name, 18, ConditionDb.greaterOrEqual)];
+    listOfUsers =  await storageService.getListOfUsers(where: filters);
+    expect(listOfUsers.length, 2);
+
+    final filters2 = [FilterDb(UserColumnName.age.name, 18, ConditionDb.less)];
+    listOfUsers =  await storageService.getListOfUsers(where: filters2);
+    expect(listOfUsers.length, 3);
   });
 }